requirements.txt
customtkinter
Pillow
keyboard
psutil
pygetwindow
opencv-python
torch
ultralytics
playwright
pyttsx3
SpeechRecognition
PyAudio
thefuzz